correct TIME_WAIT error

+++ src_server/server.cpp
@@ -20,9 +20,9 @@ void	Server::set_error_page(std::string &error)
 
 void	Server::read_body(Request &request, std::vector<char> &rest_buff)
 {
-	int 							status;
-	std::vector<char>::size_type	content_len;
-	std::vector<char>::size_type	size_buffer;
+	ssize_t status;
+	std::vector<char>::size_type content_len;
+	std::vector<char>::size_type size_buffer;
 
 	size_buffer = rest_buff.size();
 	if ((content_len = request.has_body()))
@@ -32,9 +32,9 @@ void	Server::read_body(Request &request, std::vector<char> &rest_buff)
 		{
 			syslog(LOG_DEBUG, "read body len => %lu", content_len - size_buffer);
 			if ((status = recv(request.socketfd,
-									&rest_buff[size_buffer], content_len - size_buffer, MSG_DONTWAIT)) < 0)
+							   &rest_buff[size_buffer], content_len - size_buffer, MSG_DONTWAIT)) < 0)
 			{
-				syslog(LOG_ERR, "ERROR to read body from socket %d error => %d %m",  request.socketfd, status);
+				syslog(LOG_ERR, "ERROR to read body from socket %d error => %ld %m", request.socketfd, status);
 				throw std::runtime_error("error to read body from socket");
 			}
 		}
